update koios, remove some logging

+++ include/frenzykv/types.h
@@ -16,7 +16,7 @@ using sequence_number_t = uint32_t;
 using bspan = ::std::span<::std::byte>;
 using const_bspan = ::std::span<const ::std::byte>;
 using file_id_t = uuid;
-using level_t = size_t;
+using level_t = int32_t;
 
 ::std::string_view as_string_view(const_bspan s);
