+++ util/options.cc
@@ -66,13 +66,13 @@ namespace frenzykv
 
     size_t options::allowed_level_file_number(level_t l) const noexcept
     {
-        if (l >= level_file_number.size()) return 0;
+        if (static_cast<size_t>(l) >= level_file_number.size()) return 0;
         return level_file_number[l];
     }
 
     size_t options::allowed_level_file_size(level_t l) const noexcept
     {
-        if (l >= level_file_size.size()) return 0;
+        if (static_cast<size_t>(l) >= level_file_size.size()) return 0;
         return level_file_size[l];
     }
